Change warnings to notices
See https://github.com/actions/toolkit/blob/main/packages/core/README.md#annotations. 

This helps teams that want to monitor and lower warnings - since not updating the PR title is not really a bug but just informational, changed it to a notice. That keeps it visible in the PR summary.